Update medoo.php
New features:
- max, min, count, has, get and avg support SQL joins.
- My Solution for: More conditions in other condition (#78)

AND and OR rotate automatically, but a specification of AND or OR is possible.  

follow where clauses are possible (and many more ;) ): 
PHP:[ A => B, C=>D]  
SQL: A=B AND C=D

PHP: ["AND" => [ A => B, C=>D]]  
SQL: A=B AND C=D

PHP: [ A => B, [E=>F, G=>H], C=>D] 
SQL: A=B AND (E=F OR G=H) AND C=D

- database functions are now usable: 
PHP: $where = ['password[#]' => "PASSWORD('secret')" ] :
SLQ: password = PASSWORD('secret')

The # marks functions. The value is not quoted.